# api.yaml
openapi: 3.0.3
info:
  title: Gestión de seguimiento docente
  description: |-
    API para gestionar el seguimiento docente
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
  - url: http://localhost:8070/api/v1/
...
paths:
...
  /login:
    post:
      tags:
        - usuarios
      summary: Devuelve los parámetros de un usuario
      description: Dados un email y una contraseña válidas y en base de datos, devuelve los parámetros del usuario asociado.
      operationId: authenticateUser
      requestBody:
        $ref: '#/components/requestBodies/LoginUsuario'
      responses:
        '200':
          $ref: '#/components/responses/UsuarioData'
        '422':
          description: Datos no válidos
  /usuarios:
    get:
      tags:
        - usuarios
      summary: Devuelve una lista de usuarios
      description: Devuelve la lista de los usuarios de la base de datos
      operationId: getUsuarios
      responses:
        '200':
          $ref: '#/components/responses/UsuarioListaData'

    post:
      tags:
        - usuarios
      summary: Crea un usuario en la base de datos
      description: Dados los datos de un usuario, y el usuario que intenta crear a dicho usuario, crea a un usuario en la base de datos.
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/CreateUsuario'
      responses:
        '201':
          description: Usuario creado con éxito
        '404':
          description: Usuario creador no válido
        '422':
          description: Datos no válidos
...
  /clases/{idClase}:
    get:
      tags:
        - clases
      summary: Devuelve información sobre una clase por su id
      description: Devuelve la asignatura y el grupo al que hace referencia la clase con id = {idClase}
      operationId: getClaseById
      parameters:
        - $ref: '#/components/parameters/idClase'
      responses:
        '200':
          $ref: '#/components/responses/ClaseData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Clase no encontrada
...
  /asignaturas/{idAsignatura}:
    get:
      tags:
        - asignaturas
      summary: Devuelve información de una asignatura por su id
      description: Devuelve información de la asignatura con id = {idAsignatura}
      operationId: getAsignaturaById
      parameters:
        - $ref: '#/components/parameters/idAsignatura'
      responses:
        '200':
          $ref: '#/components/responses/AsignaturaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Asignatura no encontrada
...
components:
  parameters:
...
    idClase:
      name: idClase
      in: path
      description: ID de la clase
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idAsignatura:
      name: idAsignatura
      in: path
      description: ID de la asignatura
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
  schemas:
    conId:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Id de la entidad
      required:
        - id
...
    UsuarioBasicData:
      type: object
      properties:
        email: 
          type: string
          description: Email del usuario
          example: 'grhopper@gmail.com'
    UsuarioLoginData:
      allOf:
        - $ref: '#/components/schemas/UsuarioBasicData'
        - type: object
          properties:
            password:
              type: string
              description: Contraseña del usuario
              example: "As34dnF5fsDlRPfdnsmRSDDFJAJ4JDRJEJaSD"
    UsuarioData:
      allOf:
        - $ref: '#/components/schemas/UsuarioBasicData'
        - type: object
          properties:
            nombre:
              type: string
              description: Nombre del Usuario
              example: Grace
            apellidos:
              type: string
              description: Apellidos del Usuario
              example: Hopper
            rol:
              type: string
              description: Rol del Usuario
              example: Admin
    UsuarioCreateData:
      allOf:
        - $ref: '#/components/schemas/UsuarioData'
        - $ref: '#/components/schemas/UsuarioLoginData'
        - type: object
          properties:
            creador:
              type: integer
              format: int64
              description: Id del usuario
              example: 67
...
    ListaUsuarios:
      type: object
      properties:
        usuarios:
          type: array
          items: 
            $ref: '#/components/schemas/conId'
...
    ClaseData:
      type: object
      properties:
        asignatura_id:
          type: integer
          format: int64
          description: Id de una asignatura
          example: 1
        grupo_id:
          type: integer
          format: int64
          description: Id de un grupo
          example: 1
    AsignaturaData:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la asignatura
          example: "Programación Declarativa"
        siglas:
          type: string
          description: Siglas de la asignatura
          example: "PD"
        departamento:
          type: string 
          description: Departamento que imparte esta asignatura
        periodo:
          type: integer
          enum:
            - 1
            - 2
          description: Periodo / Cuatrimestre en el que se imparte esta asignatura
...
  requestBodies:
...
    LoginUsuario:
      required: true
      description: Email y Contraseña de un usuario existente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsuarioLoginData'
    CreateUsuario:
      required: true
      description: Datos necesarios para la creación de un usuario
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsuarioCreateData'
...
  responses:
...
    UsuarioData:
      description: Datos de un usuario a devolver en un login
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsuarioData'
...
    UsuarioListaData:
      description: Lista de ids de usuarios a devolver a petición de una query
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListaUsuarios'
...
    ClaseData:
      description: Datos de una clase
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClaseData'
...
    AsignaturaData:
      description: Datos de una asignatura
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AsignaturaData'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY